+++ src/hooks/index.ts
@@ -4,6 +4,7 @@ export * from './useCalendarEvents';
 export * from './useCalendarRange';
 export * from './useColors';
 export * from './useDashboardVariables';
+export * from './useDebouncedValue';
 export * from './useEventFrames';
 export * from './useIntervalSelection';
 export * from './useLocalizer';